Fix group selection event deduplication logic
We deduplicate `GroupSelectionStarted` events by
keeping them into the `pendingGroupSelections`
cache. Duplicated events may occur due to chain
reorgs or because of the subscription monitoring
mechanism which re-fetches past events in case
something bad happens with the websocket
connection pointing to the Ethereum node.

So far, we removed current group selection from
cache just after ticket submission terminated.
We can't do that because a duplicated event can
come right after the removal and trigger a
new DKG breaking the ongoing one.

To fix the problem, we need to hold the cache
entry for the maximum DKG duration time. We
should also do it regardless of the protocol
outcome on the host node as other nodes may still
execute the protocol successfully.

Worth to flag that removing the cache entry after
hitting the DKG timeout is enough and we don't
need to monitor for a successful group
registration. A successful group registration
means we never have to trigger group selection
again for the group seed we hold in the cache so
we don't have to bother about removing the
cache entry just after the group is registered.
In contrary, when current group
selection fails, we need to remove the cache
entry once the timeout is hit to allow the node
to retry the group selection for that seed as soon
as possible.