Update dependency to coverage-pools
Coverage-pools contracts intended to be used for development of Token
Dashboard will be published to the NPM registry using under name with
`-dev` suffix. We need to update the dependencies configuration.
Note: for deployment of Dashboard on Testnet we're not using the
dependencies configured in the `package.lock`, instead we specify the
exact version we want to use when running the deployment workflow.