Update Coverage Pools Deposit page grid
The `Balance` section should display an empty state if a given user does
not have cov tokens.

We should not display the `Available to withdraw` section if a given
user does not have cov tokens.

We should display `Pending withdrawal` table only if a given user has
initiated a withdrawal.