Using particular version of cov-pool package
This change is not intended for merge to `main`. In order to easily
test cov-pools on Ropsten, the `withdrawalDelay` & `withdrawalTimeout`
parameter was changed on one of the `coverage-pools` branches
(`ropsten-test-params`) to much shorter value. Artifacts built on
modified code were pushed to the NPM registry
(`0.0.1-ropsten.4+ropsten-test-params.
5a18cc724133191c110ec9100aa9ccbfdf1e35e2` package). Now we're preparing
a workflow which will use this particular NPM package to build & deploy
Token Dashboard.