Prevent tooltips from being out of the viewport
When the tooltip will be out of the right side of the viewport we will move it
to the left.

For now it works only when the tooltip is out on the right side of viewport -
might be a good idea to do it for top, left and bottom in the future. Bu since
we have menu on the left it will probably never be cutted out from this side and
top/bottom directions are not as problematic a the directions on the x asis.