Allow for a low gas budget callbacks
The low gas budget callbacks were removed two commits ago to make sure relay
entry submit transaction is as cheap as possible, moving the
responsibility/complexity entirely to the application. But now I think it could
be too much for the application to handle...

Let's look at it from the perspective of tBTC v2.

With a callback: someone sees that the maintenance pool allows creating a new
wallet and they submit a transaction. In this transaction, sortition pool is
locked and a new relay entry is requested. Beacon produces an entry and executes
a callback on the application. The callback stores a new seed and emits an event
that is handled by clients performing sortition and generating group public key.

With no callback: someone sees that the maintenance pool allows creating a new
wallet and they submit a transaction. In this transaction, sortition pool is
locked and a new relay entry is requested. Beacon produces an entry and stores
it in the beacon contract. Off-chain clients of tBTC v2 need to monitor the
beacon contract to know when a new entry was submitted so that they can perform
sortition and start generating group public key. This version is more complex but
still possible to handle. The real complexity though begins when someone wants
to challenge the submitted DKG result: we'd need to make sure the same relay
entry is used for resolving the challenge as the one that was used by clients
to perform sortition.

It seems to me that implementing a low gas budget callback is much simpler than
handling synchronization of entry on the application side.