Remove isFloatingNumber flag from form validation
Remove unnecesarry isFloatingNumber flag from the validateAmountInRange function
used in form validation. We are removing it because it doesn't work well with
Max Amount button, when we have decimal places in the number that we don't want
to show in the form. Instead we created a new function - isNumeric - that will
check if the number is... well, numeric.