Update dashboard dependendencies to tbtc-v2 project
The `tbtc-v2` contracts intended to be used for building of Token
Dashboard will be published to the NPM registry under name with the
`-dev` suffix. We need to update the dependencies configuration in
`package.json` and `package-lock.json` dashboard files to use those
published packages.
For deployment of Dashboard on Testnet we're not using the dependencies
configured in the `package.lock`, instead we specify the exact version
we want to use when running the deployment workflow.