Fix updating available to withdraw amount
There was a bug when updaring available to withdraw amount while increasing the
existing pending withdrawal. The events returns the full amount of the new
pending withdrawal, which is value of the old withdrawal + added amount. That's
why we have to subtract it from balance instead of availableToWithdraw, because
value of the old withdrawal was already subtracted from the availableToWithdraw.