Fix displaying amount with "<"
Based on the decimals that we want to display in the amount we are defining
minimum amount to display without the "<" sign.
For example if we have token, that we want to display only 4 decimal places, the
minimal amount for this will be 0.0001, so every number below that will be
displayed as "<0.0001".