Use native libp2p topic relay
This changeset consist of two parts.

First, it gets rid of deprecated usages of
`Subscribe` methods directly on the `pubsub`
instance. Instead of that, it joins the topic,
caches the topic handle, and `Subscribe`
directly on the topic instance. As a consequence,
it passes the topic handle to the `channel`
instance and uses it for message publication.
Two helper interfaces have been also introduced
at `channel` level: `validator` and `publisher`.
Their purpose is to protect `pubsub` and `topic`
instances against incorrect usage within the
`channel` and make the `channel` API more
flexible.

The second part gets rid of the custom channel
forwarder logic and replaces it with native
libp2p topic relays introduced by:
https://github.com/libp2p/go-libp2p-pubsub/pull/317