Fix broken path filtering in dashboard-format.yaml
Filtering applied in `dashboard-detect-changes` job was creating output
called `dashboard`, whereas the `dashboard-code-format` job was executng
conditionally if (non-existing) `path-filter` output was true or if the
triggering event was different than PR. This way the
`dashboard-code-format` job was never executed for the Pull Requests.
The naming of the output in `dashboard-detect-changes` has been changed
to `path-filter`, which makes it consistent with other workflows.