Update `_displayAmount` method from `Token` class
Display amount in pattern `<...` based on the passed decimals to this
function. This commit will allow to change a default settings of the
`Token` class.

Eg.: A token displays `0` decimals by default. To display amount with 2
decimal places and display `<0.01` when the amount is less than `0.01`
pass decimals to the `displayAmount` fn(`decimals(<amount>,
<decimals>)`.