Verify contracts on Etherscan
We're adding a step verifying the migrated contracts on Etherscan. We're
using `truffle-plugin-verify` plugin. The `etherscan-verify.sh` script
contains a list of contracts which are being migrated and which should
be verified. If we'll add a new contract to the project, we'll need to
add the name of the new contract to the list in the script.

We decided to place the etherscan verification to a seperate job,
executed after the job which migrates contracts. This way the potential
verification failures will not block any other importatnt steps.