Register group when DKG result is approved
When DKG completes we need to register a group. This is a two step
process due to a fact that we need to shift the most expensive task
(that is storing a group in groups list) to the DKG submitter.
We first create a group that is not yet active. And once a result is
approved the group gots activated.