Check input paramters
We check input paramteters for the DKG process with requires.

We assert that values were set in function that require the data to set
prior to running the function. If the value is not set it may give
false-positive result e.g. due to the time calculations based on blocks
and block parameters.