Update way to load hardhat accounts for tests
We can specify named accounts in hardhat configuration file. To get them
in tests we can use either `getNamedAccounts` or `getUnnamedAccounts`
accounts.
With this change we are separating governance and deployer accounts in
tests. We don't want to run tests with deployer as zero account as this
may produce false-positive results in some cases.