Store DKG parameters in DKG library Data struct
The paramters are specific to DKG, so we want to keep them in DKG
library.
The paramters are not going to change for each run, so we don't want to
pass them for each DKG start.
We need to be sure the parameters are not updated during a DKG, so we
check if DKG is in progress before updating them.