Add function to challenge DKG result
Function can be used during challenge period to notify DKG that a
submitted result is invalid.
In case of a challenged result submission period starts from the
beginning so members are eligible to submit starting from the first
member again.

The DKG timeout is not reset though, the DKG timeout is still being
counted from the DKG start block.