Expose function to accept dkg result
Please note that we added dkg result as a parameter. It may seem not
stricly necessary for result approval in DKG library as there is just
one result and challenge period already passed. But we will need the DKG
result details to be provided as calldata as we're not storing details
of the result but only a hash. We also need to verify the calldata
against the submitted result hash.