Initial implementation of Groups
This is an initial implementation based on Random Beacon V1. We had to
introduce concept of pending groups to transfer costs of group
registration from the DKG result approver to DKG result submitter.

Groups are considered pending after a DKG result is submitted, and are
activated once a DKG result is approved. If a DKG result is challenged
we are removing the pending group.

Current implementation is quite expensive due to the group members
storing, but this is a point to further optimization.