Don't remove challenged group
It turned out that removing a group on challenge doesn't save us gas.
The operations are not returning gas to the caller, so we won't do it.
We will register pending groups and let them remain registered in the
storage but they will not be activated and remain in pending state.