Define group ID as uint64
It was proposed in https://github.com/keep-network/keep-core/pull/2637#discussion_r730866348
to change group ID from uint256 to uint64. It reduces maximum number of
groups that can be stored in group array, and we have to verify the
number of already stored groups before adding a next one, so the group
ID (which is an index of the stored group) doesn't overflow uint64.