Add function to remove pending group
A DKG result can be challenged so the pending group that was created
before should be removed in this case. The solution proposed in the original
commit c84f859d598b952502eecd7a2b21e6f9b021266e was too expensive, so we
remodeled the way groups are stored in
8953c53d7073795b39c35c17dd6f48e211357bbb.
We are poping just the latest entry from groups registry array which is
much cheaper than removing an entry that contained group data.
We're not removing data from the mapping as they can be overwritten if a
new pending group gets registered.

We're also simplifying activateGroup function to use the latest pending
group stored in the groupsRegistry.