Use group storage pointer
Instead of defining a memory struct and assigning it later to the
storage we point to the storage right away.

We're setting values for the candidate group. In case the candidate
group with a given public key hash was already registered we're
overwritting groupPubKey and members. activationTimestamp is expected to
be always 0 for candidate groups, and we don't allow adding a candidate
group that was already activated.