Clarification renames of DKG result properties
`signingMemberIndices` field has been renamed to `signingMembersIndices` to
clearly indicate indices come from multiple members and are not indices of
a single member.

`misbehaved` field been renamed to `misbehavedMembersIndices` to keep it aligned
with `signingMembersIndices` and `submitterMemberIndex`.

Documentation has been updated to clearly state all these fields should use
values from `[1, 64]` range.

This is just a rename and the change should be gas-transparent.