Validate misbehavedMembersIndices submitted in DKG result
For bitmap version it was enough to validate its length. Since we changed the
bitmap to an array of uint8 member indexes to save gas on further steps of
DKG (challenging and approving the result), we now need to validate if this
array has correct format. First, we need to ensure it has correct length just
like for bitmap version and secondly - what was added in this commit - we need
to ensure identifiers are unique. Otherwise, some sort of manipulation would be
possible by changing the order and eliminating group members in
`Groups.setGroupMembers`. The validation rule additionally enforces that
a[i-1] < a[i]. This allows for a simple validation and leaves no space for
manipulating the final group member indexes in `Group.setGroupMembers`.

Tests are not included. There are no tests for corner cases of DKG result
verification and there are no tests for misbehaved. This needs to be implemented
and is captured in the backlog. It is not ideal but I would prefer to do it all
at once later once all building blocks match each other.