Adding group status
Added notion of an expired, terminated and a stale group. This
functionality was moved from beacon-v1 with some adjustments to make
things work with a new random beacon group.
Additionaly all the group timestamps were changed to block numbers. This is for
calculation purposes and some consistency since ex. in Relay lib we use
block height.