Script to resolve stake owners from Keep TokenStaking contract
T staking contract supports existing KEEP stakes by allowing KEEP stakers
to use their stakes in T network and weights them based on KEEP<>T token
ratio. KEEP stake owner is cached in T staking contract and used to restrict
access to all `onlyOwnerOrOperator` functions. To cache KEEP staking contract
in T staking contract, we first need to resolve the owner. Resolving liquid
KEEP stake owner is easy - `TokenStaking.ownerOf` gives the answer immediately.
Resolving token grant stake owner is complicated and not possible to do on-chain
from an external contract. Keep `TokenStaking` knows the grant ID but does
not expose it externally. Going through `TokenGrant` may not always give the
right answer. Finally, we never know on-chain if we are dealing with
a `ManagedGrant` or not. Porting all this complexity to T staking contract does
not feel like the right choice.

It's time to make a tradeoff.

To solve this problem, we resolve all owners off-chain based on events and
hardcode them in `KeepStake` library. Worth noting, owner of a grant delegation
can never change in Keep `TokenStaking` contract and this is why hardcoding
should be acceptable.

Operator-owner pairs are resolved in the following way:

1. Take all Keep `TokenStaking` delegations ever.
2. Filter out undelegated operators.
3. Filter out canceled delegations.
4. Fetch grant stake information from `TokenGrant` for that operator
   to determine if we are dealing with grant delegation.
5. Fetch grantee address from Keep `TokenGrant` contract.
6. Check if we are dealing with `ManagedGrant` by looking for all created
   `ManagedGrants` and comparing their address against grantee address
   fetched from `TokenGrant` contract.