Modify method of punishment for misbehaviour
Initially we wanted to exclude the misbehaving operators by removing
them from the sortition pool for a fixed amount of time. But this posed
a problem in the following scenario:
- DKG starts, pool is locked
- Relay entry is requested
- Some members are inactive and they missed their turn to submit the
  result
On the one hand, this would require removing some members from the
sortition pool. On the other, no oerators should leave the pool when
it's locked.
As a solution it was decided that misbehaving operators will not be
removed from the pool, instead they will not be eligible to earn the
rewards for a fixed, governable amount of time.