Fix console errors
There was an error in console because whe assigned null to the prop of a
ReactElement type. We are changin the default value to a component that returns
null and thanks to that we don't need to use <OnlyIf> component anymore because
the component will not render if the prop is not provided.