Expect 33 signatures on the DKG result
Instead of expecting 48 signatures on the result we now expect just
33 which is the group threshold. Also, we expect that at least 90%
of the group members were active during the DKG. This change allows
to go down with DKG `submitDKGResult` gas costs. For 64 signatures,
the average gas cost is 1124329. For 33 signatures the average gas
cost is 815350.

Expecting more signatures than 33 does not make the protocol any more
secure. Even if we have 64 signatures supporting the result, an
adversary must not corrupt 33+1 members.