Changing the way we count active groups
Removed activeGroupsCount and replaced it with the approach we used in
v1. For this to work, a couple of additional checks were added for math
to add up.
This approach gets rid of the necessity of tracking, incrementing and
decrementing the active group counter. During the groups expiration it
might save us some gas since we are not retrieving any groups and
decrementing the counter.