Moved DKG result signature validation to challengeDkgResult
The previous version was:

In `submitDkgResult` we:
- check if t+1 signatures signing the result are valid,
- check if all operators who signed the result are in the sortition pool,
- if signatures are not valid or at least one operator is not in the sortition
  pool, the result is rejected.

Then, there is some time to challenge the result and the result is not accepted
until the challenge period is over. The sortition pool is locked for the entire
challenge period.

In `challengeDkgResult` - called by anyone only when something is wrong with the
result - we:
- call sortition pool and check if group members passed in `submitDkgResult` are
  what the sortition pool returns,
- if the expected member list does not match the actual member list, DKG result
  is rejected and all operators who signed the result are slashed.

What we do now:

In `submitDkgResult`:
- check if the submitter is in the sortition pool.

In `challengeDkgResult`:
- call sortition pool and check if group members passed in `submitDkgResult`
  are what the sortition pool expects,
- check if t+1 signatures signing the result are valid, and if they come from
  sortition pool members,
- if anything is wrong with the result, slash DKG result submitter.

The previous approach felt wrong for the number of reasons:
- It was expensive!!! More than 1M of gas for a group size of 64 members to
  `submitDkgResult`. This will only get worse for tBTC v2 and a group size of
  100 members. With the alternative approach proposed, the cost should be 600k.
- Checking if someone is in the sortition pool in `submitDkgResult` is not
  enough. One operator can have multiple signer seats, so we should check the
  stake of the operator. If they provided `S` signatures, we should check that
  their stake is `S*min_stake`. Think of the case when one operator claims they
  have all 64 member seats and we just check if that operator is in the
  sortition pool... It does not make sense, we should check if they have
  `64*min_stake` but it only makes `submitDkgResult` more expensive...
- ...At the same time, we want the minimum stake to be governable and we want to
  be able to lower and increase it. In case the minimum stake was increased, it
  affects only new operators joining the network. Someone not having the minimum
  stake can still be in the sortition pool and be selected by the sortition pool.
  Checking the minimum stake breaks in this case.
- See that we are duplicating some work: in `submitDkgResult` we check if
  members are in the sortition pool and in `challengeDkgResult` we check if they
  were selected by the sortition pool. Challenge checks their sortition pool
  membership implicitly.

In the new approach, minor censorship seems to be the only problem.
We should be fine with accepting this risk if the challenge period is long
enough and the reward for challenging is hefty. We need to move into
submit-challenge-accept model if we want to lower the cost and if miners can
censor transactions for more than 6h, we are not the only ones who have a
problem and this feels like something that needs to be solved on the blockchain
level, not in our protocol. Of course this will all be governable so if we
find a better solution in Q4, we will upgrade.