Call DKGValidator contract from DKG.challengeResult
Keeping the validation logic in a separate contract allows to use
try-catch and accept the challenge in case the validation reverted.
In practice, it should never happen but better be safe than sorry.
Submitter is expected to run a dry-run challenge before submitting
the result.

Moreover, moved the static validation to challenge phase to even
more reduce the costs of submitting DKG result.