Initial entry/DKG result submission eligibility delay set to 20 blocks
Default mining check interval in the client is 60 sec which is about
4 Ethereum blocks. 20 blocks eligibility delay is a bit more
operator-friendly leaving a space for at least 4 gas pric bumps.
This is especially important for DKG result submitter who needs
to approve the result to claim the reward.

It also means that the time before soft relay entry timeout is
hit, increased from 160 to 320 minutes.