DKG result challenge period increaset to ~48h
There is no validation in submitDkgResult function and we need to
ensure malicious DKG results are catched and eliminated no matter
what. The best approach would be validate everything in submitDkgResult
function but then, the cost of group creation would increase
significantly. 48h should be enough to report the problem even if
Ethereum network hard-forked.

x