Group creation frequency set to 5, group lifetime set to ~10 weeks
Creating group is really expensive. The total gas cost of submitDkgResult
followed by approveDkgResult is 780849 on average and 526488 minimum.

Assuming the average gas consumption, the cost in ETH is:
100Gwei gas price: 0.0780849 ETH
200Gwei gas price: 0.1561698 ETH
500Gwei gas price: 0.3904245 ETH
1000Gwei gas price: 0.780849 ETH

tBTC is expected to create a new wallet every one week and tBTC will
be the only user of the random beacon initially. It means that with
10 weeks group lifetime and group created every 5 requests, we have
a new group created every 5 weeks and every time, tBTC needs to
provide 20% of the total DKG cost.