DKG result submission reward set to 1000 T
Creating group is really expensive. The total gas cost of submitDkgResult
followed by approveDkgResult is 780849 on average and 526488 minimum.

Assuming the average gas consumption, the cost in ETH is:
100Gwei gas price: 0.0780849 ETH
200Gwei gas price: 0.1561698 ETH
500Gwei gas price: 0.3904245 ETH
1000Gwei gas price: 0.780849 ETH

Assuming 200Gwei gas price and T:ETH ratio same as KEEP:ETH ratio
today (2021-12-02), 1 ETH = 5500 KEEP, the reward in T based on
avg gas cost and 200Gwei gas price should be 858.9339 T.

Setting it to 1000 T to keep it rounded and have some safety margin to
make the reward more attractive