Sortition pool unlocking reward set to 100 T
The total cost of notifyDkgTimeout is 86883 on average and 85216
minimum.

Assuming the average gas consumption, the cost in ETH is:
100Gwei gas price: 0.0086883 ETH
200Gwei gas price: 0.0173766 ETH
500Gwei gas price: 0.0434415 ETH
1000Gwei gas price: 0.086883 ETH

Assuming 200Gwei gas price and T:ETH ratio same as KEEP:ETH ratio
today (2021-12-02), 1 ETH = 5500 KEEP, the reward in T based on
avg gas cost and 200Gwei gas price should be 95.5713 T.

Setting it to 100 T to keep it rounded and have some safety margin
to make the reward more attractive.