Updated sortition-pools dependency to 18a2a84
Commit 18a2a84 is on a feature branch changing how
SortitionPool.withdrawRewards interaface looks like. It can only
be called by the application (sortition pool owner) and allows to
pass beneficiary address, removing the need of having the
SortitionPool depend on the staking contract.

It also adds SortitionPool.updateOperatorStatusAndRewards allowing
to use different in-pool weight and rewards weight for the
operator. This will be needed for undelegating operator which
should be removed from the sortition pool but still earn rewards.