Authorization.authorizationIncreased function
Used by T staking contract to inform the random beacon that the
authorized amount for the given operator increased. Can only be
called when the sortition pool is not locked. Increases in-pool
weight and rewards weight in the pool proportionally to the
authorized stake amount immediatelly. Reverts if the sortition
pool is locked or if the authorization amount is below the
minimum.