Authorization.authorizationDecreaseRequested
Used by T staking contract to inform the random beacon that the
authorization decrease for the given operator has been regitered.
Can only be called when the sortition pool is not locked.

If the authorization is decreased to the amount higher than the
minimum authorization, in-pool weight and rewarded stake are
reduced in the sortition pool immediately so that the operator,
during the autorization decrease delay period, can not be
selected to new groups with stake authorization amount higher
than the one to which it is deauthorizing to. Authorized stake
amount remains the same until deauthorization request is
approved.

If the operator is requesting authorization decrease to zero,
in pool-weight is reduced to zero immediately and operator is
removed from the sortition pool. Rewarded stake stays at the
same level as before. Operator, during the authorization
decrease delay period, can not be selected to new groups but
it should still earn rewards based on their last authrized stake
until authorization decrease request is approved.

Overwrites pending authorization decrease request if the one
pending is for the amount higher than the minimum authorization.

Reverts if authorization decrease to 0 is pending.
Reverts if authorization decrease is requested to non-zero value
below the minimum authorization amount.
Reverts if the sortition pool is locked.