Optimizing gas cost by removing group members storage
Storing all members on chain is expensive. The idea behind this change
is to store a hash of dkg group members and emit the entire members
array upon dkg result approval.
Moreover, filtering active group members can be also done off chain to
simplify things on chain. In terms of verification, it can be done
during the challange phase on chain against all members and misbehaved
members that we can be verified beforehand against the sortition pool.