Block MM connection when Tally is injected by default
There was a bug with Tally and MetaMask - if "use Tally as default wallet"
option is tally was checked, then the app automatically connects to Tally wallet
when whe pick MetaMask, but it displays the MetaMask logo, which might be
confused for users.
We decided to block the possibility to connect to MM in such case and print
proper message, that tells user to uncheck the "use Tally as default wallet"
option if he want to connect to MetaMask.