Splitting submitRelayEntry
Split submitRelayEntry into two functions. First handles the happy path
where no slashing is involved. The submission must happen before the
soft timeout. The other function handles a scenario when a soft timeout
has passed and additionaly takes members for slashing. This split is
done to optimize gas usage for relay entry submission.

Another change was done around validation if the request is in progress.
It is strictly refactoring without changing any functionality. The reason for
this change was to avoid duplicate validation done in the relay entry
submission flow which also save us a bit of gas.