Add `dev` tag to the published npm package
This change adds an additional environment tag for packages published
during `Publish package` step. This adds a convenient reference to the
latest version of the `@keep-core/solidity-v1` package for development
environment. We also add a new condition for trigerring `NPM v1`
workflow - from now on the workflow will be executed also after changes
to the workflow get merged to the `main` branch.