Gas-optimized version of submitRelayEntry when soft timeout has not passed
In case the soft timeout has not passed yet, there is no point to calculate
slashing factor in `submitRelayEntry`. Also, because there is no slashing,
passing group members is not necessary and computing group members hash
to compare it with the value stored in groups registry is not necessary.

This allows to save gas. The optimized version costs 192,153 gas to call.

This changeset aims at making minimal changs to the existing codebase.

It is still a draft. The list of things that still need to be done include:
- fixing tests for delay factor calculations,
- adding more tests for new submitRelayEntry function,
- renaming `submitRelayEntry2` to just `submitRelayEntry` and dealing with
  overloaded calls in unit tests.