Calculating group members hash off-chain to verify on-chain
Instead of filtering out misbehaved members from the members array and
calculating a hash of the result, we can do this off-chain and pass just
the hash. However, this hash should be verifiable and the verify hash
function is added to the DKGValidator contract. Since the dkg submission
will be done more frequently compared to the dkg result challenge,
overall we should reduce gas usage and some byte code in the RandomBecon
contract.