Store just single wallet reference
WalletFactory is intended to run single wallet creation process at a
time (due to how DKG library works). It is not expected that the factory
will hold references to all created wallets. Storing the created wallet
will be a part of a next task we want to handle. A Wallet Manager is
expected to manage wallet lifecycle, so this can be a place where we
should register the wallets.