Update status of delegations and undelegations
There was a problem that undelegation did not updated its status after pending
perios is over and user had to refresh the page. This commit semi-fix that - use
does not have to refresh the page anymore but he has to go to a different
subpage to have it updated.

For it to work wihout going to other subpages we would have to compare the
pending period to currentDate every second for every undelegation which seems a
little heavy for the ui. We can improve it later when we figure out some better
solution.