Implement Wallets library
This is an initial implementation of Wallets library. The library is
used to register created wallets and allows interaction with the wallets
such as signatures requests and calculated signatures submission.

The implementation with single library holding references to all created
contract is expected to be cheaper than deploying contract clones which
was how Keep ECDSA worked in V1.