Replace walletID counter with public key hash
A group public key hash can be used as unique wallet identifier. This
change reduced gas costs (happy path):
 - approveDkgResult from 252427 to 208242
 - requestSignature from 50161 to 50424
 - submitSignature from 37173 to 35463

We still need to decide if we want to store group public key in the
Wallet or just hash is enough. If we don't store the whole key creating
a wallet will be cheaper but signature submission will be more expensive
as we would need to provide the whole public key with the signature.