Update deployment scripts to include Random Beacon
For now we deploy a stub of RandomBeacon, later we need to decide if we
want to still use a stub for local tests. And include pulling the actual
address for testnet and mainnet.

We also had to specify contract name for `DKGValidator` deployments as
the name clashes with the contract imported from random beacon.