Ignore external directory in git and prettier
An `external/npm/@keep-network/keep-core/artifacts` directory is created
every time the `threshold-network/solidity-contracts` dependency is
installed. This is due to a workaround in
`threshold-network/solidity-contracts` which allows us to use contracts
which come from different repos and share the same name. Creation of
the `external` folder is a by-product of that. We don't need to track
changes in that folder or run prettier on its files.