Update the tBTC rewards page
Adjust the tBTC rewards page to the proposal[1] that extends stakedrop
rewards for an additional 4 months. We will have two more allocations so
to simplify calculating the period of the reward we hardcoded the
`merkleRoot` in the `ExtendedECDSARewards.intervalsDates` array. After
generating the new merkle tree we must add the `merkleRoot` hash to a
given interval.

Ref:
[1]: https://forum.keep.network/t/proposal-to-extend-stakedrop-rewards-for-an-additional-4-months/351