Remove submission eligibility algorithm
As described in the [RFC-6](https://github.com/keep-network/tbtc-v2/pull/115)
we want to reduce complexity of the protocol by letting all the members
to submit a DKG result based on a handshake agreement that the operators
will perform submission in a fixed manner. No on-chain validation of the
eligibility will be performed, we relay on members good will to obey the
agreement.

Result submission period will be set as a parameter, we no longer
calculate it based on the group size and eligibility delay.

After submitting a result the submitter will have a precedence over
other members to approve the result during
`submitterPrecedencePeriodLength`.