Set dependency to a specific package of cov-pools
We need to use fixed version of `coverage-pools` package when building
Keep Dashboard, as latest versions include a commit in which we
introduced in the the `Asset Pool` contract a call to a `delegate`
function which is not present in the `KeepToken` contract used during
`AssetPool`'s deployment script. The latests package that does not
contain that commit is `1.1.0-dev.2`.

More details about the issue:
- in PR 206 we introduced `delegate` function in `ICollateralToken`
https://github.com/keep-network/coverage-pools/blob/main/contracts/interfaces/ICollateralToken.sol#L30
- in the deployment script `00_resolve_keep_token.ts`  KeepToken is usedrom the dependencies set in `package.json`, which does not have a`delegate` function
https://github.com/keep-network/coverage-pools/blob/main/deploy/00_resolve_keep_token.ts#L12
- AssetPool uses KeepToken in its deployment script:
https://github.com/keep-network/coverage-pools/blob/main/deploy/05_deploy_asset_pool.ts#L14
- KeepToken is passed to the Asset Pool constructor as a collateral
token and tries to call a `delegate` function which doesn't exist
https://github.com/keep-network/coverage-pools/blob/main/contracts/AssetPool.sol#L155