Prefix duplicated contract names with Ecdsa prefix
DKG and DKGValidator are contracts copied from random-beacon. They share
the same names in ecdsa and random-beacon which is not handled very well
by toolset we use. One of the examples is slither which reports
false-positive errors, see: https://github.com/keep-network/keep-core/runs/5196377375?check_suite_focus=true

Another example is typechain which we used to workaround with
`postbuild` script, but now can remove the workaround.

To stop wasting time on figuring out problems in external tools we just
decided to rename the contracts.