Register group on DKG result approval
We used to register a candidate group on DKG result submission and later
activating in on approval. It added a lot of complexity with candidates
group registration, popping the candidate group on DKG challenge etc.

Here we simplyfy the solution to register an active group only when DKG
result got approved.