Shift wallet owner initialization to post deployment
It is expected that the Wallet Owner will be a Bridge contract that will
be deployed after the WalletRegistry. We can't expect the Wallet Owner's
address in the constructor. We defined a function that allows initial
configuration of Wallet Owner's address without a governance delay. Any
subsequent update to the wallet will require governance delay to pass.